+++ src/app/signup/owner/page.test.tsx
@@ -16,7 +16,7 @@ import {
     submitSignup,
 } from "~tests/utils/form-helpers";
 import OwnerSignup from "./page";
-import { OwnersAPI } from "src/api/owners-api";
+import { OwnersAPI } from "src/api/ownersAPI";
 
 /**
  * Test suites and mock functions generated with GPT-5 mini and help from:
