+++ src/models/owner.ts
